+++ Util/CustomReturnUrlUtil.php
@@ -17,8 +17,11 @@ declare(strict_types=1);
 
 namespace MultiSafepay\ConnectCore\Util;
 
+use Magento\Framework\App\ObjectManager;
+use Magento\Framework\App\ProductMetadataInterface;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\UrlInterface;
+use Magento\Quote\Model\QuoteIdToMaskedQuoteIdInterface;
 use Magento\Sales\Api\Data\OrderInterface;
 use Magento\Store\Api\StoreRepositoryInterface;
 use MultiSafepay\ConnectCore\Config\Config;
@@ -48,23 +51,32 @@ class CustomReturnUrlUtil
     private $logger;
 
     /**
+     * @var QuoteIdToMaskedQuoteIdInterface|null
+     */
+    private $quoteIdToMaskedQuoteId;
+
+    /**
      * CustomReturnUrlUtil constructor.
      *
      * @param Config $config
      * @param SecureToken $secureToken
      * @param StoreRepositoryInterface $storeRepository
      * @param Logger $logger
+     * @param ProductMetadataInterface $productMetadata
      */
     public function __construct(
         Config $config,
         SecureToken $secureToken,
         StoreRepositoryInterface $storeRepository,
-        Logger $logger
+        Logger $logger,
+        ProductMetadataInterface $productMetadata
     ) {
         $this->config = $config;
         $this->secureToken = $secureToken;
         $this->storeRepository = $storeRepository;
         $this->logger = $logger;
+        $this->quoteIdToMaskedQuoteId = version_compare($productMetadata->getVersion(), '2.3', '>=')
+            ? ObjectManager::getInstance()->get(QuoteIdToMaskedQuoteIdInterface::class) : null;
     }
 
     /**
@@ -125,6 +137,7 @@ class CustomReturnUrlUtil
         $availableCustomVariables = [
             '{{order.increment_id}}' => $order->getIncrementId(),
             '{{order.order_id}}' => $order->getEntityId(),
+            '{{quote.quote_id}}' => $order->getQuoteId(),
             '{{payment.code}}' => $order->getPayment()->getMethod(),
             '{{payment.transaction_id}}' => $transactionParameters['transactionid'] ?? '',
             '{{store.unsecure_base_url}}' => $orderStore->getBaseUrl(UrlInterface::URL_TYPE_WEB),
@@ -134,6 +147,11 @@ class CustomReturnUrlUtil
             '{{secure_token}}' => $transactionParameters['secureToken'] ?? $this->secureToken->generate((string)
                 $order->getRealOrderId()),
         ];
+
+        if ($this->quoteIdToMaskedQuoteId) {
+            $availableCustomVariables['{{quote.masked_id}}'] =
+                $this->quoteIdToMaskedQuoteId->execute((int)$order->getQuoteId());
+        }
 
         foreach ($availableCustomVariables as $var => $value) {
             $urlString = str_replace($var, $value, $urlString);